// convex/_generated/api.ts
/* eslint-disable */
/**
 * STUB FILE - This file will be auto-generated when you run `npx convex dev`
 *
 * This stub allows TypeScript compilation to succeed before Convex is initialized.
 * DO NOT edit this file manually - it will be overwritten.
 */

export const api = {
  companies: {
    listCompaniesWithStats: {} as any,
    getCompanyWithContacts: {} as any,
    getCompany: {} as any,
    searchCompanies: {} as any,
    listCompanies: {} as any,
    createCompany: {} as any,
    updateCompany: {} as any,
    deleteCompany: {} as any,
  },
  contacts: {
    listContacts: {} as any,
    getContact: {} as any,
    searchContacts: {} as any,
    findContactByPhone: {} as any,
    getContactWithCalls: {} as any,
    createContact: {} as any,
    createContactFromCall: {} as any,
    updateContact: {} as any,
    deleteContact: {} as any,
    addContactNote: {} as any,
  },
  calls: {
    listCallsForInbox: {} as any,
    getCallDetail: {} as any,
    getRecentCalls: {} as any,
    getUnmatchedCallsCount: {} as any,
    getCallStats: {} as any,
    linkCallToContact: {} as any,
    unlinkCallFromContact: {} as any,
    deleteCall: {} as any,
  },
  sync: {
    triggerManualSync: {} as any,
    processCDRBatchFromAPI: {} as any,
    completeSync: {} as any,
    failSync: {} as any,
  },
};

export const internal = {
  sync: {
    getLastSync: {} as any,
    getRunningSyncs: {} as any,
    getSyncHistory: {} as any,
    createSyncRecord: {} as any,
    completeSyncRecord: {} as any,
    failSyncRecord: {} as any,
    processCDRBatch: {} as any,
    syncYeastarCDR: {} as any,
  },
};